requirements.txt
PyQt6
requests
psycopg2-binary
bcrypt
